requirements.txt
flask>=3.0.0
anthropic>=0.25.0
python-docx>=1.1.0
docx2pdf>=0.1.8
requests>=2.31.0
